removed all codes from main_window
