+++ angrmanagement/ui/main_window.py
@@ -3,19 +3,18 @@ import logging
 import sys
 from typing import Optional
 
-from PySide2.QtWidgets import QMainWindow, QTabWidget, QFileDialog, QProgressBar, QDialog, QLabel, QLineEdit, \
-    QPushButton, QVBoxLayout
+from PySide2.QtWidgets import QMainWindow, QTabWidget, QFileDialog, QProgressBar
 from PySide2.QtWidgets import QMessageBox, QSplitter, QShortcut
 from PySide2.QtGui import QResizeEvent, QIcon, QDesktopServices, QKeySequence
 from PySide2.QtCore import Qt, QSize, QEvent, QTimer, QUrl
 
 import angr
+
 try:
     from angr.angrdb import AngrDB
 except ImportError as ex:
     print(str(ex))
     AngrDB = None  # type: Optional[type]
-
 
 try:
     import archr
@@ -56,49 +55,11 @@ class MainWindow(QMainWindow):
 
     username = None
 
-    class Modal(QDialog):
-        def __init__(self, outerClass):
-            super().__init__(outerClass)
-            self.username = None
-            # Create widgets
-            self.label = QLabel("Enter your Name")
-            self.edit = QLineEdit("")
-            self.button = QPushButton("Submit")
-            self.outerClass = outerClass
-
-            # Create layout and add widgets
-
-            layout = QVBoxLayout()
-
-            layout.addWidget(self.label)
-            layout.addWidget(self.edit)
-            layout.addWidget(self.button)
-
-            # Set dialog layout
-
-            self.setLayout(layout)
-
-            # Add button signal to greetings slot
-
-            self.button.clicked.connect(self.greetings)
-
-        # Greets the user
-
-        def greetings(self):
-            self.outerClass.username = self.edit.text()
-            self.close()
-
-        def closeEvent(self, event):
-            if not self.outerClass.username:
-                event.ignore()
-
-
     def __init__(self, parent=None, show=True):
         super(MainWindow, self).__init__(parent)
 
         icon_location = os.path.join(IMG_LOCATION, 'angr.png')
         self.setWindowIcon(QIcon(icon_location))
-        self._fatigue_flag = True
 
         GlobalInfo.main_window = self
 
@@ -137,8 +98,6 @@ class MainWindow(QMainWindow):
 
         # I'm ready to show off!
         if show:
-            self.modal = self.Modal(self)
-            self.modal.show()
             self.showMaximized()
             self.windowHandle().screenChanged.connect(self.on_screen_changed)
             self.show()
@@ -361,8 +320,6 @@ class MainWindow(QMainWindow):
 
         for plugin in list(self.workspace.plugins.active_plugins):
             self.workspace.plugins.deactivate_plugin(plugin)
-        self._fatigue_flag = False
-        self.modal.close()
         event.accept()
 
     def event(self, event):
@@ -402,10 +359,10 @@ class MainWindow(QMainWindow):
 
     def open_docker_button(self):
         required = {
-            'archr: git clone https://github.com/angr/archr && cd archr && pip install -e .':archr,
-            'keystone: pip install --no-binary keystone-engine keystone-engine':keystone
-            }
-        is_missing = [ key for key, value in required.items() if value is None ]
+            'archr: git clone https://github.com/angr/archr && cd archr && pip install -e .': archr,
+            'keystone: pip install --no-binary keystone-engine keystone-engine': keystone
+        }
+        is_missing = [key for key, value in required.items() if value is None]
         if len(is_missing) > 0:
             req_msg = 'You need to install the following:\n\n\t' + '\n\t'.join(is_missing)
             req_msg += '\n\nInstall them to enable this functionality.'
@@ -519,7 +476,7 @@ class MainWindow(QMainWindow):
     def run_induction_variable_analysis(self):
         self.workspace.view_manager.first_view_in_category('disassembly').run_induction_variable_analysis()
 
-    def run_dependency_analysis(self, func_addr: Optional[int]=None, func_arg_idx: Optional[int]=None):
+    def run_dependency_analysis(self, func_addr: Optional[int] = None, func_arg_idx: Optional[int] = None):
         if self.workspace is None or self.workspace.instance is None:
             return
         dep_analysis_job = DependencyAnalysisJob(func_addr=func_addr, func_arg_idx=func_arg_idx)
@@ -607,7 +564,7 @@ class MainWindow(QMainWindow):
 
     def _recalculate_view_sizes(self, old_size):
         adjustable_dockable_views = [dock for dock in self.workspace.view_manager.docks
-                                     if dock.widget().default_docking_position in ('left', 'bottom', )]
+                                     if dock.widget().default_docking_position in ('left', 'bottom',)]
 
         if not adjustable_dockable_views:
             return
